[fix] Fix unacceptable letters in input problem
the save project dialog would lose focus on random letters being
pressed, this pr fixes that by moving the ExportAnnotationsDialog out of
the MenuItems being mapped on, to only have a single instance
